+++ server/src/providers/foldingRange.ts
@@ -1,4 +1,4 @@
-import { FoldingRange, FoldingRangeKind, FoldingRangeParams, Position, Range } from 'vscode-languageserver/node';
+import { FoldingRange, FoldingRangeKind, FoldingRangeParams, Range } from 'vscode-languageserver/node';
 import { documents } from '../utils/variables';
 import * as ld from '../utils/languageDefinitions';
 import { getParsedDocument } from '../utils/functions';
@@ -11,13 +11,13 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 	if (doc === undefined) {return null;}
 	const parsed = await getParsedDocument(params.textDocument.uri);
 	if (parsed === undefined) {return null;}
-	var result: FoldingRange[] = [];
+	const result: FoldingRange[] = [];
 
-	var openranges: FoldingRange[] = [];
-	var inmultilinemacro: boolean = false;
-	var dotteddolevel: number = 0;
-	var injsonxdata: boolean = false;
-	var routinename = "";
+	const openranges: FoldingRange[] = [];
+	let inMultiLineMacro: boolean = false;
+	let dottedDoLevel: number = 0;
+	let inJSONXData: boolean = false;
+	let routinename = "";
 	for (let line = 0; line < parsed.length; line++) {
 		if (parsed[line].length === 0) {
 			if (openranges.length > 0 && openranges[openranges.length-1].kind === FoldingRangeKind.Comment) {
@@ -33,7 +33,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 					openranges[idx].endLine = line-1;
 					result.push(openranges[idx]);
 					openranges.splice(idx,1);
-					dotteddolevel--;
+					dottedDoLevel--;
 				}
 			}
 			continue;
@@ -69,13 +69,13 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 				}
 				openranges.pop();
 			}
-			if (inmultilinemacro) {
+			if (inMultiLineMacro) {
 				// Check if the last token is a ##Continue
 				if (
 					parsed[line][parsed[line].length-1].l !== ld.cos_langindex || parsed[line][parsed[line].length-1].s !== ld.cos_ppf_attrindex ||
 					doc.getText(Range.create(
-						Position.create(line,parsed[line][parsed[line].length-1].p),
-						Position.create(line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c)
+						line,parsed[line][parsed[line].length-1].p,
+						line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c
 					)).toLowerCase() !== "continue"
 				) {
 					// This is the end of a multi-line macro
@@ -91,21 +91,21 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 						result.push(openranges[prevrange]);
 					}
 					openranges.splice(prevrange,1);
-					inmultilinemacro = false;
+					inMultiLineMacro = false;
 				}
 			}
 			if (
 				parsed[line][parsed[line].length-1].l == ld.cls_langindex && parsed[line][parsed[line].length-1].s == ld.cls_delim_attrindex &&
 				doc.getText(Range.create(
-					Position.create(line,parsed[line][parsed[line].length-1].p),
-					Position.create(line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c)
-				)) === "{"
+					line,parsed[line][parsed[line].length-1].p,
+					line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c
+				)) == "{"
 			) {
 				// This line ends with a UDL open curly
 
 				if (
 					(parsed[line].length === 1 && parsed[line-1][0].l == ld.cls_langindex && parsed[line-1][0].s == ld.cls_keyword_attrindex &&
-					doc.getText(Range.create(Position.create(line-1,parsed[line-1][0].p),Position.create(line-1,parsed[line-1][0].p+parsed[line-1][0].c))).toLowerCase() === "class")
+					doc.getText(Range.create(line-1,parsed[line-1][0].p,line-1,parsed[line-1][0].p+parsed[line-1][0].c)).toLowerCase() === "class")
 					||
 					(parsed[line].length > 1 && parsed[line][0].l == ld.cls_langindex && parsed[line][0].s == ld.cls_keyword_attrindex &&
 					firsttokentext.toLowerCase() === "class")
@@ -131,16 +131,16 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 						(parsed[nl][0].s === ld.cls_keyword_attrindex ||
 						(parsed[nl][0].s === ld.cls_delim_attrindex && 
 						doc.getText(Range.create(
-							Position.create(nl,parsed[nl][0].p),
-							Position.create(nl,parsed[nl][0].p+parsed[nl][0].c)
+							nl,parsed[nl][0].p,
+							nl,parsed[nl][0].p+parsed[nl][0].c
 						)) === "}"))
 					) {
 						// Close the member range
 						if (
 							parsed[nl][0].s === ld.cls_delim_attrindex && 
 							doc.getText(Range.create(
-								Position.create(nl,parsed[nl][0].p),
-								Position.create(nl,parsed[nl][0].p+parsed[nl][0].c)
+								nl,parsed[nl][0].p,
+								nl,parsed[nl][0].p+parsed[nl][0].c
 							)) === "}"
 						) {
 							openranges[openranges.length-1].endLine = nl-1;
@@ -166,7 +166,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 				var foundopencurly =  false;
 				for (let tkn = 1; tkn < parsed[line].length; tkn++) {
 					if (parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_brace_attrindex) {
-						const bracetext = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+						const bracetext = doc.getText(Range.create(line,parsed[line][tkn].p,line,parsed[line][tkn].p+parsed[line][tkn].c));
 						if (bracetext === "{") {
 							foundopencurly = true;
 							break;
@@ -222,7 +222,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 				firsttokentext.toLowerCase() === "routine"
 			) {
 				// This is the ROUTINE header line
-				routinename = doc.getText(Range.create(Position.create(line,parsed[line][1].p),Position.create(line,parsed[line][1].p+parsed[line][1].c)));
+				routinename = doc.getText(Range.create(line,parsed[line][1].p,line,parsed[line][1].p+parsed[line][1].c));
 			}
 			if (
 				parsed[line].length >= 2 &&
@@ -231,7 +231,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 			) {
 				// This line starts with a COS preprocessor command
 
-				const ppc = doc.getText(Range.create(Position.create(line,parsed[line][0].p),Position.create(line,parsed[line][1].p+parsed[line][1].c))).toLowerCase();
+				const ppc = doc.getText(Range.create(line,parsed[line][0].p,line,parsed[line][1].p+parsed[line][1].c)).toLowerCase();
 				if (ppc === "#if" || ppc === "#ifdef" || ppc === "#ifndef" || ppc === "#ifundef") {
 					// These preprocessor commands always open a new range
 					openranges.push({
@@ -280,8 +280,8 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 					if (
 						parsed[line][parsed[line].length-1].l == ld.cos_langindex && parsed[line][parsed[line].length-1].s == ld.cos_ppf_attrindex &&
 						doc.getText(Range.create(
-							Position.create(line,parsed[line][parsed[line].length-1].p),
-							Position.create(line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c)
+							line,parsed[line][parsed[line].length-1].p,
+							line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c
 						)).toLowerCase() === "continue"
 					) {
 						// This is the start of a multi-line macro definition
@@ -290,7 +290,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 							endLine: line,
 							kind: "isc-mlmacro"
 						});
-						inmultilinemacro = true;
+						inMultiLineMacro = true;
 					}
 				}
 			}
@@ -302,8 +302,8 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 					if (parsed[line][xmltkn].l == ld.xml_langindex && parsed[line][xmltkn].s == ld.xml_tagdelim_attrindex) {
 						// This is a tag delimiter 
 						const tokentext = doc.getText(Range.create(
-							Position.create(line,parsed[line][xmltkn].p),
-							Position.create(line,parsed[line][xmltkn].p+parsed[line][xmltkn].c)
+							line,parsed[line][xmltkn].p,
+							line,parsed[line][xmltkn].p+parsed[line][xmltkn].c
 						));
 						if (tokentext === "<") {
 							// Open a new XML range
@@ -334,7 +334,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 			else if (parsed[line].length > 1 && parsed[line][0].l == ld.cls_langindex && parsed[line][0].s == ld.cls_delim_attrindex) {
 				// This line starts with a UDL delimiter
 
-				const firsttwochars = doc.getText(Range.create(Position.create(line,0),Position.create(line,2)));
+				const firsttwochars = doc.getText(Range.create(line,0,line,2));
 				if (firsttwochars === "</") {
 					// Close the most recent Storage range
 					var prevrange = openranges.length-1;
@@ -359,7 +359,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 							stkn !== parsed[line].length - 1 &&
 							parsed[line][stkn].l == ld.cls_langindex && parsed[line][stkn].s == ld.cls_delim_attrindex &&
 							parsed[line][stkn+1].l == ld.cls_langindex && parsed[line][stkn+1].s == ld.cls_delim_attrindex &&
-							doc.getText(Range.create(Position.create(line,parsed[line][stkn].p),Position.create(line,parsed[line][stkn+1].p+parsed[line][stkn+1].c))) === "</"
+							doc.getText(Range.create(line,parsed[line][stkn].p,line,parsed[line][stkn+1].p+parsed[line][stkn+1].c)) === "</"
 						) {
 							closed = 1;
 							break;
@@ -423,20 +423,20 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 						if (parsed[line][k].l == ld.cls_langindex && parsed[line][k].s == ld.cls_keyword_attrindex) {
 							// This is a UDL trailing keyword
 							const keytext = doc.getText(Range.create(
-								Position.create(line,parsed[line][k].p),
-								Position.create(line,parsed[line][k].p+parsed[line][k].c)
+								line,parsed[line][k].p,
+								line,parsed[line][k].p+parsed[line][k].c
 							)).toLowerCase();
 							if (keytext === "mimetype") {
 								// The MimeType keyword is present
 								if (parsed[line][k+2] !== undefined) {
 									// An MimeType is specified
 									const mimetype = doc.getText(Range.create(
-										Position.create(line,parsed[line][k+2].p+1),
-										Position.create(line,parsed[line][k+2].p+parsed[line][k+2].c-1)
+										line,parsed[line][k+2].p+1,
+										line,parsed[line][k+2].p+parsed[line][k+2].c-1
 									));
 									if (mimetype === "application/json") {
 										// This is the start of an XData block containing JSON
-										injsonxdata = true;
+										inJSONXData = true;
 									}
 								}
 								break;
@@ -444,18 +444,18 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 						}
 					}
 				}
-				else if (injsonxdata && keytext !== "xdata") {
+				else if (inJSONXData && keytext !== "xdata") {
 					// We've reached the next class member
-					injsonxdata = false;
+					inJSONXData = false;
 				}
 			}
-			else if (injsonxdata) {
+			else if (inJSONXData) {
 				// We're in a JSON XData block so look for opening/closing curly braces and brackets
 
 				for (let tkn = 0; tkn < parsed[line].length; tkn++) {
 					if (parsed[line][tkn].l === ld.javascript_langindex && parsed[line][tkn].s === ld.javascript_delim_attrindex) {
 						// This is a JSON bracket
-						const jb = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+						const jb = doc.getText(Range.create(line,parsed[line][tkn].p,line,parsed[line][tkn].p+parsed[line][tkn].c));
 						if (jb === "[" || jb === "{") {
 							// Create a new JSON range
 							openranges.push({
@@ -486,7 +486,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 				for (let tkn = 0; tkn < parsed[line].length; tkn++) {
 					if (parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_jsonb_attrindex) {
 						// This is a JSON bracket
-						const jb = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+						const jb = doc.getText(Range.create(line,parsed[line][tkn].p,line,parsed[line][tkn].p+parsed[line][tkn].c));
 						if (jb === "[" || jb === "{") {
 							// Create a new JSON range
 							openranges.push({
@@ -511,20 +511,20 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 							openranges.splice(prevrange,1);
 						}
 					}
-					if (tkn+1 > dotteddolevel && parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_dots_attrindex) {
+					if (tkn+1 > dottedDoLevel && parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_dots_attrindex) {
 						// This is the start of a dotted Do
-						dotteddolevel++;
+						dottedDoLevel++;
 						openranges.push({
 							startLine: line-1,
 							endLine: line-1,
 							kind: "isc-dotteddo"
 						});
 					}
-					if (tkn === 0 && dotteddolevel > 0) {
+					if (tkn === 0 && dottedDoLevel > 0) {
 						// We're in a dotted Do, so check if the line begins with the correct number of dots
 
-						if (parsed[line].length >= dotteddolevel) {
-							for (let level = dotteddolevel-1; level >= 0; level--) {
+						if (parsed[line].length >= dottedDoLevel) {
+							for (let level = dottedDoLevel-1; level >= 0; level--) {
 								if (parsed[line][level].l !== ld.cos_langindex || parsed[line][level].s !== ld.cos_dots_attrindex) {
 									// This dotted Do level is closed
 									var prevrange = openranges.length-1;
@@ -537,13 +537,13 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 									openranges[prevrange].endLine = line-1;
 									result.push(openranges[prevrange]);
 									openranges.splice(prevrange,1);
-									dotteddolevel--;
+									dottedDoLevel--;
 								}
 							}
 						}
 						else {
 							// At least one dotted Do level is closed
-							for (let level = dotteddolevel-1; level >= 0; level--) {
+							for (let level = dottedDoLevel-1; level >= 0; level--) {
 								if (level > parsed[line].length-1) {
 									// Close all dotted Do levels that are greater than the length of this line
 									var prevrange = openranges.length-1;
@@ -556,7 +556,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 									openranges[prevrange].endLine = line-1;
 									result.push(openranges[prevrange]);
 									openranges.splice(prevrange,1);
-									dotteddolevel--;
+									dottedDoLevel--;
 								}
 								else if (parsed[line][level].l !== ld.cos_langindex || parsed[line][level].s !== ld.cos_dots_attrindex) {
 									// This dotted Do level is closed
@@ -570,7 +570,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 									openranges[prevrange].endLine = line-1;
 									result.push(openranges[prevrange]);
 									openranges.splice(prevrange,1);
-									dotteddolevel--;
+									dottedDoLevel--;
 								}
 							}
 						}
@@ -603,7 +603,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 			// Done with special processing, so loop again to find all ObjectScript braces, UDL parentheses and HTML script tags
 			for (let tkn = 0; tkn < parsed[line].length; tkn++) {
 				if (parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_brace_attrindex) {
-					const bracetext = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+					const bracetext = doc.getText(Range.create(line,parsed[line][tkn].p,line,parsed[line][tkn].p+parsed[line][tkn].c));
 					if (bracetext === "{") {
 						// Open a new ObjectScript code block range
 						openranges.push({
@@ -629,7 +629,7 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 					}
 				}
 				else if (parsed[line][tkn].l === ld.cls_langindex && parsed[line][tkn].s === ld.cls_delim_attrindex) {
-					const delimtext = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+					const delimtext = doc.getText(Range.create(line,parsed[line][tkn].p,line,parsed[line][tkn].p+parsed[line][tkn].c));
 					if (delimtext === "(") {
 						// Open a new UDL parentheses range
 						openranges.push({
@@ -659,12 +659,12 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 					parsed[line][tkn].l == ld.html_langindex && parsed[line][tkn].s == ld.html_delim_attrindex &&
 					parsed[line][tkn+1].l == ld.html_langindex && parsed[line][tkn+1].s == ld.html_tag_attrindex &&
 					doc.getText(Range.create(
-						Position.create(line,parsed[line][tkn].p),
-						Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)
+						line,parsed[line][tkn].p,
+						line,parsed[line][tkn].p+parsed[line][tkn].c
 					)) === "<" &&
 					doc.getText(Range.create(
-						Position.create(line,parsed[line][tkn+1].p),
-						Position.create(line,parsed[line][tkn+1].p+parsed[line][tkn+1].c)
+						line,parsed[line][tkn+1].p,
+						line,parsed[line][tkn+1].p+parsed[line][tkn+1].c
 					)).toLowerCase() === "script"
 				) {
 					// Open a new HTML script tag range
@@ -680,8 +680,8 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 					parsed[line][tkn+1].l == ld.html_langindex && parsed[line][tkn+1].s == ld.html_delim_attrindex &&
 					parsed[line][tkn+2].l == ld.html_langindex && parsed[line][tkn+2].s == ld.html_tag_attrindex &&
 					doc.getText(Range.create(
-						Position.create(line,parsed[line][tkn+2].p),
-						Position.create(line,parsed[line][tkn+2].p+parsed[line][tkn+2].c)
+						line,parsed[line][tkn+2].p,
+						line,parsed[line][tkn+2].p+parsed[line][tkn+2].c
 					)).toLowerCase() === "script"
 				) {
 					// Close the most recent HTML script tag range
@@ -697,6 +697,23 @@ export async function onFoldingRanges(params: FoldingRangeParams) {
 						result.push(openranges[prevrange]);
 					}
 					openranges.splice(prevrange,1);
+				}
+				else if (parsed[line][tkn].l == ld.cos_langindex && parsed[line][tkn].s == ld.cos_comment_attrindex) {
+					const inCComment = openranges.length && openranges[openranges.length - 1].kind == "isc-ccomment";
+					if (!inCComment && doc.getText(Range.create(line,parsed[line][tkn].p,line,parsed[line][tkn].p+2)) == "/*") {
+						// Open a new C-style comment range
+						openranges.push({
+							startLine: line,
+							endLine: line,
+							kind: "isc-ccomment"
+						});
+					} else if (inCComment && doc.getText(Range.create(line,parsed[line][tkn].p+parsed[line][tkn].c-2,line,parsed[line][tkn].p+parsed[line][tkn].c)) == "*/") {
+						// Close the most recent C-style comment range
+						const cCommentRange = openranges.pop();
+						cCommentRange.endLine = line - 1;
+						cCommentRange.kind = FoldingRangeKind.Comment;
+						if (cCommentRange.endLine > cCommentRange.startLine) result.push(cCommentRange);
+					}
 				}
 			}
 		}